master/requirements.txt
regex
discord.py
asyncio
yt_dlp
requests
pynacl